crypto::cipher: publicise traits for external use
This makes `Tls12AeadAlgorithm` and `Tls13AeadAlgorithm` public, as well as
the types that are associated with them.

Document fields that need to become public to allow `Tls12CipherSuite` and
`Tls13CipherSuite` to become public.
